+++ tests/e2e/core-tests/specs/shopper/front-end-cart.test.js
@@ -3,8 +3,8 @@
  * Internal dependencies
  */
 const {
-	CustomerFlow,
-	StoreOwnerFlow,
+	shopper,
+	merchant,
 	createSimpleProduct,
 	uiUnblocked
 } = require( '@woocommerce/e2e-utils' );
@@ -21,58 +21,58 @@ const {
 const runCartPageTest = () => {
 	describe('Cart page', () => {
 		beforeAll(async () => {
-			await StoreOwnerFlow.login();
+			await merchant.login();
 			await createSimpleProduct();
-			await StoreOwnerFlow.logout();
+			await merchant.logout();
 		});
 
 		it('should display no item in the cart', async () => {
-			await CustomerFlow.goToCart();
+			await shopper.goToCart();
 			await expect(page).toMatchElement('.cart-empty', {text: 'Your cart is currently empty.'});
 		});
 
 		it('should add the product to the cart when "Add to cart" is clicked', async () => {
-			await CustomerFlow.goToShop();
-			await CustomerFlow.addToCartFromShopPage('Simple product');
+			await shopper.goToShop();
+			await shopper.addToCartFromShopPage('Simple product');
 
-			await CustomerFlow.goToCart();
-			await CustomerFlow.productIsInCart('Simple product');
+			await shopper.goToCart();
+			await shopper.productIsInCart('Simple product');
 		});
 
 		it('should increase item qty when "Add to cart" of the same product is clicked', async () => {
-			await CustomerFlow.goToShop();
-			await CustomerFlow.addToCartFromShopPage('Simple product');
+			await shopper.goToShop();
+			await shopper.addToCartFromShopPage('Simple product');
 
-			await CustomerFlow.goToCart();
-			await CustomerFlow.productIsInCart('Simple product', 2);
+			await shopper.goToCart();
+			await shopper.productIsInCart('Simple product', 2);
 		});
 
 		it('should update qty when updated via qty input', async () => {
-			await CustomerFlow.goToCart();
-			await CustomerFlow.setCartQuantity('Simple product', 4);
+			await shopper.goToCart();
+			await shopper.setCartQuantity('Simple product', 4);
 			await expect(page).toClick('button', {text: 'Update cart'});
 			await uiUnblocked();
 
-			await CustomerFlow.productIsInCart('Simple product', 4);
+			await shopper.productIsInCart('Simple product', 4);
 		});
 
 		it('should remove the item from the cart when remove is clicked', async () => {
-			await CustomerFlow.goToCart();
-			await CustomerFlow.removeFromCart('Simple product');
+			await shopper.goToCart();
+			await shopper.removeFromCart('Simple product');
 			await uiUnblocked();
 
 			await expect(page).toMatchElement('.cart-empty', {text: 'Your cart is currently empty.'});
 		});
 
 		it('should update subtotal in cart totals when adding product to the cart', async () => {
-			await CustomerFlow.goToShop();
-			await CustomerFlow.addToCartFromShopPage('Simple product');
+			await shopper.goToShop();
+			await shopper.addToCartFromShopPage('Simple product');
 
-			await CustomerFlow.goToCart();
-			await CustomerFlow.productIsInCart('Simple product', 1);
+			await shopper.goToCart();
+			await shopper.productIsInCart('Simple product', 1);
 			await expect(page).toMatchElement('.cart-subtotal .amount', {text: '$9.99'});
 
-			await CustomerFlow.setCartQuantity('Simple product', 2);
+			await shopper.setCartQuantity('Simple product', 2);
 			await expect(page).toClick('button', {text: 'Update cart'});
 			await uiUnblocked();
 
@@ -80,7 +80,7 @@ const runCartPageTest = () => {
 		});
 
 		it('should go to the checkout page when "Proceed to Checkout" is clicked', async () => {
-			await CustomerFlow.goToCart();
+			await shopper.goToCart();
 			await Promise.all([
 				page.waitForNavigation({waitUntil: 'networkidle0'}),
 				expect(page).toClick('.checkout-button', {text: 'Proceed to checkout'}),
