+++ tests/e2e/core-tests/specs/shopper/front-end-checkout.test.js
@@ -3,8 +3,8 @@
  * Internal dependencies
  */
 const {
-	CustomerFlow,
-	StoreOwnerFlow,
+	shopper,
+	merchant,
 	createSimpleProduct,
 	setCheckbox,
 	settingsPageSaveChanges,
@@ -19,11 +19,11 @@ let orderId;
 const runCheckoutPageTest = () => {
 	describe('Checkout page', () => {
 		beforeAll(async () => {
-			await StoreOwnerFlow.login();
+			await merchant.login();
 			await createSimpleProduct();
 
 			// Go to general settings page
-			await StoreOwnerFlow.openSettings('general');
+			await merchant.openSettings('general');
 
 			// Set base location with state CA.
 			await expect(page).toSelect('select[name="woocommerce_default_country"]', 'United States (US) — California');
@@ -44,7 +44,7 @@ const runCheckoutPageTest = () => {
 			]);
 
 			// Enable BACS payment method
-			await StoreOwnerFlow.openSettings('checkout', 'bacs');
+			await merchant.openSettings('checkout', 'bacs');
 			await setCheckbox('#woocommerce_bacs_enabled');
 			await settingsPageSaveChanges();
 
@@ -52,7 +52,7 @@ const runCheckoutPageTest = () => {
 			await verifyCheckboxIsSet('#woocommerce_bacs_enabled');
 
 			// Enable COD payment method
-			await StoreOwnerFlow.openSettings('checkout', 'cod');
+			await merchant.openSettings('checkout', 'cod');
 			await setCheckbox('#woocommerce_cod_enabled');
 			await settingsPageSaveChanges();
 
@@ -60,28 +60,28 @@ const runCheckoutPageTest = () => {
 			await verifyCheckboxIsSet('#woocommerce_cod_enabled');
 
 			// Enable PayPal payment method
-			await StoreOwnerFlow.openSettings('checkout', 'paypal');
+			await merchant.openSettings('checkout', 'paypal');
 			await setCheckbox('#woocommerce_paypal_enabled');
 			await settingsPageSaveChanges();
 
 			// Verify that settings have been saved
 			await verifyCheckboxIsSet('#woocommerce_paypal_enabled');
 
-			await StoreOwnerFlow.logout();
+			await merchant.logout();
 		});
 
 		it('should display cart items in order review', async () => {
-			await CustomerFlow.goToShop();
-			await CustomerFlow.addToCartFromShopPage(simpleProductName);
-			await CustomerFlow.goToCheckout();
-			await CustomerFlow.productIsInCheckout(simpleProductName, `1`, `9.99`, `9.99`);
+			await shopper.goToShop();
+			await shopper.addToCartFromShopPage(simpleProductName);
+			await shopper.goToCheckout();
+			await shopper.productIsInCheckout(simpleProductName, `1`, `9.99`, `9.99`);
 		});
 
 		it('allows customer to choose available payment methods', async () => {
-			await CustomerFlow.goToShop();
-			await CustomerFlow.addToCartFromShopPage(simpleProductName);
-			await CustomerFlow.goToCheckout();
-			await CustomerFlow.productIsInCheckout(simpleProductName, `2`, `19.98`, `19.98`);
+			await shopper.goToShop();
+			await shopper.addToCartFromShopPage(simpleProductName);
+			await shopper.goToCheckout();
+			await shopper.productIsInCheckout(simpleProductName, `2`, `19.98`, `19.98`);
 
 			await expect(page).toClick('.wc_payment_method label', {text: 'PayPal'});
 			await expect(page).toClick('.wc_payment_method label', {text: 'Direct bank transfer'});
@@ -89,18 +89,18 @@ const runCheckoutPageTest = () => {
 		});
 
 		it('allows customer to fill billing details', async () => {
-			await CustomerFlow.goToShop();
-			await CustomerFlow.addToCartFromShopPage(simpleProductName);
-			await CustomerFlow.goToCheckout();
-			await CustomerFlow.productIsInCheckout(simpleProductName, `3`, `29.97`, `29.97`);
-			await CustomerFlow.fillBillingDetails(config.get('addresses.customer.billing'));
+			await shopper.goToShop();
+			await shopper.addToCartFromShopPage(simpleProductName);
+			await shopper.goToCheckout();
+			await shopper.productIsInCheckout(simpleProductName, `3`, `29.97`, `29.97`);
+			await shopper.fillBillingDetails(config.get('addresses.customer.billing'));
 		});
 
 		it('allows customer to fill shipping details', async () => {
-			await CustomerFlow.goToShop();
-			await CustomerFlow.addToCartFromShopPage(simpleProductName);
-			await CustomerFlow.goToCheckout();
-			await CustomerFlow.productIsInCheckout(simpleProductName, `4`, `39.96`, `39.96`);
+			await shopper.goToShop();
+			await shopper.addToCartFromShopPage(simpleProductName);
+			await shopper.goToCheckout();
+			await shopper.productIsInCheckout(simpleProductName, `4`, `39.96`, `39.96`);
 
 			// Select checkbox to ship to a different address
 			await page.evaluate(() => {
@@ -108,22 +108,22 @@ const runCheckoutPageTest = () => {
 			});
 			await uiUnblocked();
 
-			await CustomerFlow.fillShippingDetails(config.get('addresses.customer.shipping'));
+			await shopper.fillShippingDetails(config.get('addresses.customer.shipping'));
 		});
 
 		it('allows guest customer to place order', async () => {
-			await CustomerFlow.goToShop();
-			await CustomerFlow.addToCartFromShopPage(simpleProductName);
-			await CustomerFlow.goToCheckout();
-			await CustomerFlow.productIsInCheckout(simpleProductName, `5`, `49.95`, `49.95`);
-			await CustomerFlow.fillBillingDetails(config.get('addresses.customer.billing'));
+			await shopper.goToShop();
+			await shopper.addToCartFromShopPage(simpleProductName);
+			await shopper.goToCheckout();
+			await shopper.productIsInCheckout(simpleProductName, `5`, `49.95`, `49.95`);
+			await shopper.fillBillingDetails(config.get('addresses.customer.billing'));
 
 			await uiUnblocked();
 
 			await expect(page).toClick('.wc_payment_method label', {text: 'Cash on delivery'});
 			await expect(page).toMatchElement('.payment_method_cod', {text: 'Pay with cash upon delivery.'});
 			await uiUnblocked();
-			await CustomerFlow.placeOrder();
+			await shopper.placeOrder();
 
 			await expect(page).toMatch('Order received');
 
@@ -134,8 +134,8 @@ const runCheckoutPageTest = () => {
 		});
 
 		it('store owner can confirm the order was received', async () => {
-			await StoreOwnerFlow.login();
-			await StoreOwnerFlow.openAllOrdersView();
+			await merchant.login();
+			await merchant.openAllOrdersView();
 
 			// Click on the order which was placed in the previous step
 			await Promise.all([
