+++ tests/e2e/utils/package.json
@@ -11,6 +11,7 @@
   "main": "build/index.js",
   "module": "build-module/index.js",
   "dependencies": {
+    "@wordpress/deprecated": "^2.10.0",
     "@wordpress/e2e-test-utils": "^4.6.0",
     "config": "3.3.3",
     "faker": "^5.1.0",
