+++ tests/e2e/utils/src/flows/index.js
@@ -1,0 +1,14 @@
+/**
+ * Internal dependencies
+ */
+const flowConstants = require( './constants' );
+const flowExpressions = require( './expressions' );
+const merchant = require( './merchant' );
+const shopper = require( './shopper' );
+
+module.exports = {
+	...flowConstants,
+	...flowExpressions,
+	merchant,
+	shopper,
+};
