+++ tests/e2e/utils/src/index.js
@@ -6,6 +6,7 @@ export * from '@wordpress/e2e-test-utils';
  * Internal dependencies
  */
 export * from './flows';
+export * from './old-flows';
 export * from './components';
 export * from './page-utils';
